Remove useless inner data in HostWrapper
